refactor: Add cursor-pointer class to Image component in AKB page
This commit adds the "cursor-pointer" class to the Image component in the AKB page. The class is added to the Image component's className prop, allowing the image to have a cursor pointer when hovered over. This refactor improves the user experience by providing visual feedback to users that the image is clickable.

Note: The commit message has been generated based on the provided code changes and recent repository commits.