refactor: Remove onClick event from Image component in AKB page
This commit removes the onClick event from the Image component in the AKB page. The onClick event was previously used to show a PopUp modal with the clicked image, but it has been removed as part of the code changes. This refactor simplifies the code and removes unnecessary functionality.

Note: The commit message has been generated based on the provided code changes and recent repository commits.