refactor: Add useOutsideAlerter hook to PopUp component
This commit adds the useOutsideAlerter hook to the PopUp component in the AKB page. The hook is used to detect clicks outside of the PopUp component and close it if necessary. This refactor improves the user experience by allowing users to close the PopUp by clicking outside of it.

Note: The commit message has been generated based on the provided code changes and recent repository commits.